requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
numpy>=1.25.0